conectada base de datos y preparado para lanzar

+++ src/routes/routes.js
@@ -47,7 +47,7 @@ routes.post("/login", (req, resp) => {
       if (passpoint == hashcode(obj_points, varphi)) {
         const accessToken = logger.get_accesstoken({
           email,
-          usuario: results[0].usuario,
+          username: results[0].usuario,
         });
         resp.send({
           accesstoken: accessToken,
@@ -84,20 +84,22 @@ routes.post("/singup", uploads.single("image"), async (req, resp) => {
     });
     return;
   }
-  let result = req.file.path;
-  try {
-    result = (await cloudinary.v2.uploader.upload(req.file.path)).url;
-    fs.unlinkSync(req.file.path);
-  } catch (err) {
-    console.log(err);
-    resp.send({
-      accesstoken: undefined,
-      error: "hubo un error guardando la imagen",
-    });
-    fs.unlinkSync(req.file.path);
-    return;
+  let result = "";
+
+  if (_default == "false") {
+    try {
+      result = (await cloudinary.v2.uploader.upload(req.file.path)).url;
+      fs.unlinkSync(req.file.path);
+    } catch (err) {
+      resp.send({
+        accesstoken: undefined,
+        error: "hubo un error guardando la imagen",
+      });
+      fs.unlinkSync(req.file.path);
+      return;
+    }
   }
-  const imagepath = (_default == "true") ? "/images/default.jpg" : result;
+  const imagepath = (_default == "true") ? process.env.DEF_IMG : result;
   connect.query(
     `SELECT correo
   FROM Usuario WHERE correo='${email}';`,
@@ -135,4 +137,57 @@ routes.get("/tolerance", (req, res) => {
   res.send({ tolerance });
 });
 
+routes.post("/update/:at", uploads.single("image"), (req, res) => {
+  const { at } = req.params;
+  const { points, confirmation, _default } = req.body;
+  const obj_points = JSON.parse(points);
+  const obj_conf = JSON.parse(confirmation);
+  const varphi = getVarphi(obj_points);
+  const passcode = hashcode(obj_points, varphi);
+  const lit_varphi = JSON.stringify(varphi);
+  const { email } = logger.verify(at);
+
+  if (passcode != hashcode(obj_conf, varphi)) {
+    res.send({
+      accesstoken: undefined,
+      error: "las contrasenas no coinciden",
+    });
+    return;
+  }
+
+  connect.query(
+    `SELECT imagen FROM Usuario WHERE correo='${email}'`,
+    async (err, resp, dat) => {
+      let imagen = resp[0].imagen;
+      console.log(imagen);
+      if (_default == "false") {
+        const elim = await cloudinary.v2.uploader.destroy(getPublicId(imagen));
+        imagen = (await cloudinary.v2.uploader.upload(req.file.path)).url;
+      }
+      connect.query(
+        `UPDATE Usuario SET passpoint='${passcode}', phi='${lit_varphi}',imagen='${imagen}' WHERE correo='${email}';`,
+        (err, resp, dat) => {
+          if (err) {
+            res.send({
+              accesstoken: undefined,
+              error: "hubo un error al guardar la contrasena",
+            });
+            return;
+          }
+          res.send({
+            accesstoken: at,
+            error: "hubo un error al guardar la contrasena",
+          });
+        },
+      );
+    },
+  );
+});
+
+function getPublicId(url) {
+  const parts = url.split("/");
+  const filename = parts[parts.length - 1];
+  const publicId = filename.split(".")[0];
+  return publicId;
+}
 module.exports = routes;
